refactor: 스크롤 네비게이터 기능 커스텀 훅으로 분리
