+++ src/pages/Main.tsx
@@ -6,37 +6,16 @@ import Banner from '../components/Banner';
 import styled from '@emotion/styled';
 import { SCENE_CARDS } from '../components/sceneCard/consts';
 import ResponsiveFlexGrid from '../components/common/ResponsiveFlexGrid';
-import { useRef } from 'react';
+import { useScrollNavigation } from '../hooks/useScrollNavigation';
 
 const Main = () => {
-  const scenesRef = useRef<HTMLDivElement>(null);
-  const contributorsRef = useRef<HTMLDivElement>(null);
-
-  const HEADER_HEIGHT = 91;
-
-  const scrollToScenes = () => {
-    if (scenesRef.current) {
-      const elementTop = scenesRef.current.offsetTop;
-      window.scrollTo({
-        top: elementTop - HEADER_HEIGHT,
-        behavior: 'smooth',
-      });
-    }
-  };
-
-  const scrollToContributors = () => {
-    if (contributorsRef.current) {
-      const elementTop = contributorsRef.current.offsetTop;
-      window.scrollTo({
-        top: elementTop - HEADER_HEIGHT,
-        behavior: 'smooth',
-      });
-    }
-  };
-
-  const scrollToTop = () => {
-    window.scrollTo({ top: 0, behavior: 'smooth' });
-  };
+  const {
+    scenesRef,
+    contributorsRef,
+    scrollToScenes,
+    scrollToContributors,
+    scrollToTop,
+  } = useScrollNavigation();
 
   return (
     <>
@@ -67,6 +46,7 @@ const Main = () => {
     </>
   );
 };
+
 export default Main;
 
 const Container = styled.div`
